+++ tests/unit/components/paper-sidenav-toggle-test.js
@@ -1,4 +1,3 @@
-import Ember from 'ember';
 import {
   moduleForComponent,
   test
@@ -20,42 +19,5 @@ test('it renders', function(assert) {
   // appends the component to the page
   this.render();
   assert.equal(component._state, 'inDOM');
-  assert.equal(component.get('element').tagName, 'md-sidenav-toggle'.toUpperCase(), 'matches `md-sidenav-toggle`');
-});
-
-test('when clicked triggers `toggleSidenav` on navContainer', function(assert) {
-  assert.expect(1);
-  var component = this.subject();
-
-  var didFire = false;
-  component.set('navContainer', Ember.Object.create({
-    toggleSidenav: function() {
-      didFire = true;
-    }
-  }));
-
-  var $component = this.$();
-  Ember.run(function() {
-    $component.trigger('click');
-  });
-
-  assert.ok(didFire, 'toggleSidenav called');
-});
-
-test('when clicked triggers `expandSidenav` if toggle is false on navContainer', function(assert) {
-  assert.expect(1);
-  var component = this.subject({toggle: false});
-  var didFire = false;
-  component.set('navContainer', Ember.Object.create({
-    expandSidenav: function() {
-      didFire = true;
-    }
-  }));
-
-  var $component = this.$();
-  Ember.run(function() {
-    $component.trigger('click');
-  });
-
-  assert.ok(didFire, 'expandSidenav called');
+  assert.equal(component.get('element').tagName, 'div'.toUpperCase(), 'matches `div`');
 });